+++ src/wt5001/wt5001.h
@@ -149,7 +149,7 @@ namespace upm {
      * @param len Length of the buffer
      * @return Number of bytes read
      */
-    int readData(char *buffer, size_t len);
+    int readData(char *buffer, int len);
 
     /**
      * Writes the data in the buffer to the device
@@ -158,7 +158,7 @@ namespace upm {
      * @param len Length of the buffer
      * @return Number of bytes written
      */
-    int writeData(char *buffer, size_t len);
+    int writeData(char *buffer, int len);
 
     /**
      * Sets up proper tty I/O modes and the baud rate. The default
